Added max replica

+++ src/Application/Vagrant/Services/VagrantService.cs
@@ -325,6 +325,7 @@ public class VagrantService(ILogger<VagrantService> logger)
                   config.vm.network "public_network", bridge: "Default Switch"
                   config.vm.provider "hyperv" do |hv|
                     hv.enable_virtualization_extensions = true
+                    hv.linked_clone = true
                     hv.memory = "{1024 * memoryGB}"
                     hv.cpus = "{cpus}"
                   end
@@ -484,6 +485,11 @@ public class VagrantService(ILogger<VagrantService> logger)
         {
             try
             {
+                if (!dir.DirectoryExists())
+                {
+                    break;
+                }
+
                 await DeleteVMCore(dir, id, cancellationToken.WithTimeout(TimeSpan.FromMinutes(2)));
 
                 await WaitKill(dir, cancellationToken.WithTimeout(TimeSpan.FromMinutes(2)));
